+++ api/web/src/main.js
@@ -94,10 +94,6 @@ const router = new VueRouter.createRouter({
                     name: 'home-menu-datas',
                     component: () => import('./components/CloudTAK/Menu/Datas.vue')
                 },{
-                    path: 'datas/user',
-                    name: 'home-menu-datas-user',
-                    component: () => import('./components/CloudTAK/Menu/DataUser.vue')
-                },{
                     path: 'contacts',
                     name: 'home-menu-contacts',
                     component: () => import('./components/CloudTAK/Menu/Contacts.vue')
